pyproject.toml
[project]
name = "geminilivedemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.64.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.1",
]
